Moved miscellaneous utilities to their own file under utilities
