+++ src/components/SlidePreview.vue
@@ -7,7 +7,7 @@
 /* tslint:enable */
 import { Vue, Component, Watch, Prop } from "vue-property-decorator";
 import GraphicModel from "../models/GraphicModel";
-import Utilities from "../utilities";
+import Utilities from "../utilities/miscellaneous";
 import * as SVG from "svg.js";
 
 @Component
