+++ src/components/StyleEditor.vue
@@ -12,7 +12,7 @@
 <script lang="ts">
 /* tslint:enable */
 import { Vue, Component, Prop, Watch } from "vue-property-decorator";
-import Utilities from "../utilities";
+import Utilities from "../utilities/miscellaneous";
 import StyleModel from "../models/StyleModel";
 
 @Component
