+++ src/store.ts
@@ -1,7 +1,7 @@
 import Vue from "vue";
 import Vuex from "vuex";
 import SlideModel from "./models/SlideModel";
-import Utilities from "./utilities";
+import Utilities from "./utilities/miscellaneous";
 import Tools from "./utilities/tools";
 import Theme from "./models/Theme";
 import GraphicModel from "./models/GraphicModel";
